fix ads

+++ MovieWebsite-backend/src/main/java/com/project/MovieWebsite/services/impl/RateServiceImpl.java
@@ -2,9 +2,8 @@ package com.project.MovieWebsite.services.impl;
 
 import com.project.MovieWebsite.dtos.RateDTO;
 import com.project.MovieWebsite.exceptions.DataNotFoundException;
-import com.project.MovieWebsite.models.Movie;
-import com.project.MovieWebsite.models.Rate;
-import com.project.MovieWebsite.models.User;
+import com.project.MovieWebsite.models.*;
+import com.project.MovieWebsite.repositories.ManagerStorageRateRepository;
 import com.project.MovieWebsite.repositories.MovieRepository;
 import com.project.MovieWebsite.repositories.RateRepository;
 import com.project.MovieWebsite.repositories.UserRepository;
@@ -26,7 +25,7 @@ public class RateServiceImpl implements RateService {
     private final MovieRepository movieRepository;
     private final UserRepository userRepository;
     private final RateRepository rateRepository;
-
+    private final ManagerStorageRateRepository managerStorageRateRepository;
 
     @Override
     public Rate createRate(RateDTO rateDTO) throws DataNotFoundException {
@@ -56,8 +55,11 @@ public class RateServiceImpl implements RateService {
     }
 
     @Override
-    public Rate getRate(int id) {
-        return rateRepository.findById(id).orElseThrow(() -> new RuntimeException("Rate not found"));
+    public Movie getRateByMovie(int movieId) throws DataNotFoundException{
+        Movie existingMovie= movieRepository.findById(movieId)
+                .orElseThrow(() -> new DataNotFoundException("Cannot find movie with id: "+movieId));
+
+        return existingMovie;
     }
 
     @Override
@@ -86,7 +88,13 @@ public class RateServiceImpl implements RateService {
 
     @Override
     public void deleteRateMonth() {
-        LocalDate cutoffDate = LocalDate.now().minusMonths(1);
-        rateRepository.deleteRateOlderThanMonth(cutoffDate);
+        ManagerStorageRate managerStorageRate= managerStorageRateRepository.getAll();
+        if(managerStorageRate!=null){
+            managerStorageRate.setLastDelete(LocalDate.now());
+            managerStorageRateRepository.save(managerStorageRate);
+            LocalDate cutoffDate = LocalDate.now().minusMonths(1);
+            rateRepository.deleteRateOlderThanMonth(cutoffDate);
+        }
+
     }
 }
